+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### [13.30.1](https://kaos.sh/ek/13.30.1)
+
+- **`[hashutil]`** Added method `Hash.Equal`
+- **`[hashutil]`** Added method `Hash.EqualString`
+- **`[hashutil]`** Added method `Hash.IsEmpty`
+
 ### [13.30.0](https://kaos.sh/ek/13.30.0)
 
 - **`[options]`** Added method `Map.SetIf`